requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.16
django-filter==24.3
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
httpie==3.2.2
idna==3.7
jinxed==1.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==2.2.0
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
psycopg2-binary==2.9.10
Pygments==2.18.0
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==3.5.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
